Add ActionMenu and PauseMenu scenes
I have added the ActionMenu and PauseMenu as 2D scenes initially, then converted them to CanvasLayers in order to have the menus stay at the center of the screen rather than one part of the scene they are added to.

I have also added empty buttons to the two scenes.